Move data related providers to data module
