+++ app/src/main/java/com/aconno/sensorics/dagger/application/AppModule.kt
@@ -298,54 +298,4 @@ class AppModule {
     fun provideTimeProvider(): TimeProvider {
         return TimeProviderImpl()
     }
-
-    @Provides
-    @Singleton
-    fun provideStoreDataUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): StoreDataUseCase {
-        return StoreDataUseCase(FileStorageImpl(sensoricsApplication.applicationContext))
-    }
-
-    @Provides
-    @Singleton
-    fun provideStoreTextUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): StoreTextUseCase {
-        return StoreTextUseCase(StoreDataUseCase(
-            FileStorageImpl(sensoricsApplication.applicationContext)))
-    }
-
-    @Provides
-    @Singleton
-    fun provideReadDataUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): ReadDataUseCase {
-        return ReadDataUseCase(FileStorageImpl(sensoricsApplication.applicationContext))
-    }
-
-    @Provides
-    @Singleton
-    fun provideReadTextUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): ReadTextUseCase {
-        return ReadTextUseCase(ReadDataUseCase(FileStorageImpl(sensoricsApplication.applicationContext)))
-    }
-
-    @Provides
-    @Singleton
-    fun provideStoreTempDataUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): StoreTempDataUseCase {
-        return StoreTempDataUseCase(FileStorageImpl(sensoricsApplication.applicationContext))
-    }
-
-    @Provides
-    @Singleton
-    fun provideStoreTempTextUseCase(
-        sensoricsApplication: SensoricsApplication
-    ): StoreTempTextUseCase {
-        return StoreTempTextUseCase(StoreTempDataUseCase(
-            FileStorageImpl(sensoricsApplication.applicationContext)))
-    }
 }
